+++ db/src/main/java/net/bytle/db/fs/FsTableSystem.java
@@ -264,8 +264,8 @@ public class FsTableSystem extends TableSystem {
   }
 
   @Override
-  public Integer size(DataPath dataPath) {
-    int i = 0;
+  public long size(DataPath dataPath) {
+    long i = 0;
     try (SelectStream selectStream = getSelectStream(dataPath)) {
       while (selectStream.next()) {
         i++;
